Organized file structure, removed props from constructor
