+++ src/App/App.scss
@@ -1,4 +1,4 @@
-@import 'variables.scss';
+@import '../Styles/variables.scss';
 
 .app-render {
   text-align: center;
@@ -33,4 +33,3 @@
   text-align: left;
   text-decoration: none;
   }
-
