+++ src/Card/Card.scss
@@ -1,4 +1,4 @@
-@import 'variables.scss';
+@import '../Styles/variables.scss';
 
 .card {
   display: inline-block;
