+++ src/CompareContainer/CompareContainer.css
