+++ src/CompareContainer/CompareContainer.scss
