+++ src/CompareContainer/CompareContainer.test.js
