+++ src/ComparedCard/ComparedCard.js
