+++ src/ComparedCard/ComparedCard.scss
