+++ src/ComparedCard/ComparedCard.test.js
