+++ src/Controls/Controls.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const Controls = () => {
+  return (
+
+  );
+};
