--- src/Styles/index.scss
@@ -1,14 +1,0 @@
-@import 'Card.scss';
-@import 'App.scss';
-@import 'CardContainer.scss';
-@import 'CompareContainer.scss';
-@import 'DataButton.scss';
-
-* {
-  margin: 0;
-  padding: 0;
-  font-family: sans-serif;
-  box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  -webkit-box-sizing: border-box;
-}
