+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
-import './Styles/index.css';
+import './index/index.css';
 
-import App from './App';
+import App from './App/App';
 
 render( <App />, document.getElementById('root') );
