+++ src/index/index.scss
@@ -1,0 +1,13 @@
+@import '../Card/Card.scss';
+@import '../App/App.scss';
+@import '../CardContainer/CardContainer.scss';
+@import '../CompareContainer/CompareContainer.scss';
+
+* {
+  margin: 0;
+  padding: 0;
+  font-family: sans-serif;
+  box-sizing: border-box;
+  -moz-box-sizing: border-box;
+  -webkit-box-sizing: border-box;
+}
